% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{water_pal}
\alias{water_pal}
\title{A California Water Boards palette generator}
\usage{
water_pal(name, n, type = c("discrete", "continuous"))
}
\arguments{
\item{name}{Name of desired palette. Choices are:
\code{bardo}, \code{cwqmc},  \code{cwqmcdark}, \code{tribalbanner}, 
\code{tribalheader},  \code{wbregions}, \code{datasalmon}, 
\code{logogradient}, \code{drought}}

\item{n}{Number of colors desired. Unfortunately most palettes now only
have 4 or 5 colors. F. Color schemes are derived from the raw_images 
folder in this package, with the help of mycolorpicker and mycolor.space.
If omitted, uses all colours.}

\item{type}{Either "continuous" or "discrete". Use continuous if you want
to automatically interpolate between colours.
@importFrom graphics rgb rect par image text}
}
\value{
A vector of colours.
}
\description{
These are color palettes based on art and iconography from the California 
State Water Resources Control Board.
}
\examples{
water_pal("cwqmc")
water_pal("tribalbanner")
water_pal("wbregions")
water_pal("wbregions", 3)

# If you need more colours than normally found in a palette, you
# can use a continuous palette to interpolate between existing
# colours
pal <- water_pal(21, name = "bardo", type = "continuous")
image(volcano, col = pal)
}
\keyword{colors}
