% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\docType{data}
\name{water_palettes}
\alias{water_palettes}
\title{Complete list of palettes}
\format{
An object of class \code{list} of length 9.
}
\usage{
water_palettes
}
\description{
Use \code{\link{water_pal}} to construct palettes of desired length.
}
\keyword{datasets}
