% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waterpal.R
\docType{package}
\name{waterpal}
\alias{waterpal}
\title{waterpal}
\description{
list of palettes for use by the California Water Boards, USA
}
\details{
list of palettes for use by the California Water Boards, USA
}
